refactor(admin/people): fusionner les details syndic/agence/garant en une seule route
Les trois pages de detail dediees /admin/people/{syndics,agencies,guarantors}/[id]
creaient des doublons presque identiques. Elles sont remplacees par une route unifiee
/admin/people/[id] qui dispatche sur le role via profiles.role et delegue aux sections
specialisees. Les routes historiques owners/tenants/vendors sont preservees (UI riche
existante) via redirection.
