+++ app/admin/people/[id]/error.tsx
@@ -1,5 +1,5 @@
 "use client";
 import { DashboardError } from "@/components/ui/dashboard-error";
 export default function ErrorPage({ error, reset }: { error: Error & { digest?: string }; reset: () => void }) {
-  return <DashboardError error={error} reset={reset} section="admin" title="Erreur - Détail garant" returnHref="/admin/people?tab=guarantors" />;
+  return <DashboardError error={error} reset={reset} section="admin" title="Erreur - Détail utilisateur" returnHref="/admin/people" />;
 }
