+++ app/admin/people/[id]/loading.tsx
