--- app/admin/people/guarantors/[id]/loading.tsx
@@ -1,4 +1,0 @@
-import DashboardLoading from "@/components/ui/dashboard-loading";
-export default function Loading() {
-  return <DashboardLoading variant="detail" />;
-}
