--- app/admin/people/syndics/[id]/error.tsx
@@ -1,5 +1,0 @@
-"use client";
-import { DashboardError } from "@/components/ui/dashboard-error";
-export default function ErrorPage({ error, reset }: { error: Error & { digest?: string }; reset: () => void }) {
-  return <DashboardError error={error} reset={reset} section="admin" title="Erreur - Détail syndic" returnHref="/admin/people?tab=syndics" />;
-}
